Drop support for Node.js 10
Update dependencies

Move tests to __tests__ and exclude from package
