+++ .gitignore
@@ -1,4 +1,5 @@
 ### Node ###
+*.tgz
 
 # Logs
 logs
