+++ changelog.md
@@ -1,13 +1,28 @@
-# `@carnesen/redirector` change log
-## 0.1.2
-### Fixed
-Fix bug where Koa middleware was not properly calling next().
-## 0.1.1
+# **@carnesen/redirector** changelog
+
+## 0.2.0 (2020-08-05)
+
 ### Breaking
+
+- Drop support for Node.js 10
+
+## 0.1.2
+
+### Fixed
+
+Fix bug where Koa middleware was not properly calling next().
+
+## 0.1.1
+
+### Breaking
+
 - Drop support for Node.js 8.
+
 ### Added
+
 - Convert to TypeScript
 - Support Node.js 12, 14
 
 ## 0.0.0
+
 Initial release
