+++ src/__tests__/start.test.ts
@@ -1,0 +1,12 @@
+import { start } from '../start';
+
+describe(__filename, () => {
+	it('starts and stops', async () => {
+		const server = start();
+		await new Promise((resolve, reject) => {
+			server.on('listening', resolve);
+			server.on('error', reject);
+		});
+		server.close();
+	});
+});
