+++ src/index.ts
@@ -1,7 +1,7 @@
 import { start } from './start';
 
 if (require.main === module) {
-  start();
+	start();
 }
 
 export { start };
