+++ src/start.ts
@@ -5,8 +5,8 @@ import { echo } from './util';
 const PORT = Number(process.env.PORT) || 8000;
 
 export function start(port = PORT): Server {
-  echo(`Starting server`);
-  return app.listen(port, () => {
-    echo(`Listening on port ${port}`);
-  });
+	echo(`Starting server`);
+	return app.listen(port, () => {
+		echo(`Listening on port ${port}`);
+	});
 }
