+++ src/util.ts
@@ -1,3 +1,3 @@
 export function echo(...args: Parameters<typeof console.log>): void {
-  console.log(...args); // eslint-disable-line no-console
+	console.log(...args); // eslint-disable-line no-console
 }
